% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_dsc.R
\name{theme_dsc}
\alias{theme_dsc}
\title{ggplot DSC-WAV theme}
\usage{
theme_dsc(size = 12, family = "Oswald")
}
\description{
This function is a theme function for use with ggplot2.
It tries to have a very tight composition, suitable for email or chat message
}
\details{
http://ggplot2.tidyverse.org/reference/theme.html
}
